@@ -1,10 +1,16 @@
 module github.com/fergusstrange/embedded-postgres
 
-go 1.13
+go 1.18
 
 require (
 	github.com/lib/pq v1.10.4
 	github.com/stretchr/testify v1.7.0
 	github.com/xi2/xz v0.0.0-20171230120015-48954b6210f8
 	go.uber.org/goleak v1.1.12
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
