@@ -2,6 +2,13 @@ module github.com/fergusstrange/embedded-postgres/platform-test
 
 replace github.com/fergusstrange/embedded-postgres => ../
 
-go 1.13
+go 1.18
 
 require github.com/fergusstrange/embedded-postgres v0.0.0
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/lib/pq v1.10.4 // indirect
+	github.com/xi2/xz v0.0.0-20171230120015-48954b6210f8 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
