@@ -8,7 +8,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-//nolint:funlen
 func Test_DefaultVersionStrategy_AllGolangDistributions(t *testing.T) {
 	allGolangDistributions := map[string][]string{
 		"aix/ppc64":       {"aix", "ppc64"},
@@ -54,28 +53,42 @@ func Test_DefaultVersionStrategy_AllGolangDistributions(t *testing.T) {
 		"windows/arm":     {"windows", "arm"},
 	}
 
+	versionDifferences := map[PostgresVersion]map[string][]string{
+		PostgresVersion("14.0.0"): {},
+		PostgresVersion("14.1.0"): {},
+		PostgresVersion("14.2.0"): {"darwin/arm64": {"darwin", "arm64v8"}},
+		V14:                       {"darwin/arm64": {"darwin", "arm64v8"}},
+	}
 	defaultConfig := DefaultConfig()
 
-	for dist, expected := range allGolangDistributions {
-		dist := dist
-		expected := expected
+	for version, differences := range versionDifferences {
+		defaultConfig.version = version
 
-		t.Run(fmt.Sprintf("DefaultVersionStrategy_%s", dist), func(t *testing.T) {
-			osArch := strings.Split(dist, "/")
+		for dist, expected := range allGolangDistributions {
+			dist := dist
+			expected := expected
 
-			operatingSystem, architecture, postgresVersion := defaultVersionStrategy(
-				defaultConfig,
-				osArch[0],
-				osArch[1],
-				linuxMachineName,
-				func() bool {
-					return false
-				})()
+			if override, ok := differences[dist]; ok {
+				expected = override
+			}
 
-			assert.Equal(t, expected[0], operatingSystem)
-			assert.Equal(t, expected[1], architecture)
-			assert.Equal(t, V12, postgresVersion)
-		})
+			t.Run(fmt.Sprintf("DefaultVersionStrategy_%s", dist), func(t *testing.T) {
+				osArch := strings.Split(dist, "/")
+
+				operatingSystem, architecture, postgresVersion := defaultVersionStrategy(
+					defaultConfig,
+					osArch[0],
+					osArch[1],
+					linuxMachineName,
+					func() bool {
+						return false
+					})()
+
+				assert.Equal(t, expected[0], operatingSystem)
+				assert.Equal(t, expected[1], architecture)
+				assert.Equal(t, version, postgresVersion)
+			})
+		}
 	}
 }
 
@@ -92,7 +105,7 @@ func Test_DefaultVersionStrategy_Linux_ARM32V6(t *testing.T) {
 
 	assert.Equal(t, "linux", operatingSystem)
 	assert.Equal(t, "arm32v6", architecture)
-	assert.Equal(t, V12, postgresVersion)
+	assert.Equal(t, V14, postgresVersion)
 }
 
 func Test_DefaultVersionStrategy_Linux_ARM32V7(t *testing.T) {
@@ -108,7 +121,7 @@ func Test_DefaultVersionStrategy_Linux_ARM32V7(t *testing.T) {
 
 	assert.Equal(t, "linux", operatingSystem)
 	assert.Equal(t, "arm32v7", architecture)
-	assert.Equal(t, V12, postgresVersion)
+	assert.Equal(t, V14, postgresVersion)
 }
 
 func Test_DefaultVersionStrategy_Linux_Alpine(t *testing.T) {
@@ -126,7 +139,7 @@ func Test_DefaultVersionStrategy_Linux_Alpine(t *testing.T) {
 
 	assert.Equal(t, "linux", operatingSystem)
 	assert.Equal(t, "amd64-alpine", architecture)
-	assert.Equal(t, V12, postgresVersion)
+	assert.Equal(t, V14, postgresVersion)
 }
 
 func Test_DefaultVersionStrategy_shouldUseAlpineLinuxBuild(t *testing.T) {
